requirements-generic.txt
Babel
beautifulsoup4
certifi
charset-normalizer
colorama
config
datedelta
Django
djangorestframework
doc2pdf
gusregon
idna
imgkit
iniconfig
isodate
Jinja2
litex
litex.regon
lxml
Markdown
MarkupSafe
migen
mypy
numpy
packaging
pandas
pdfkit
pillow
platformdirs
pluggy
psycopg-binary
psycopg2-binary
pylint
pyserial
pytesseract
pytest
pytest-django
pytest-env
python-dateutil
python-docx
python-dotenv
python-stdnum
pytz
PyVirtualDisplay
requests
requests-file
requests-toolbelt
setuptools
six
soupsieve
sqlparse
tqdm
typing_extensions
tzdata
urllib3
zeep
